fix typo in Regression Models chapter
Dear Dr. Clark,

First of all, I would like to thank you the wonderful book “Bayesian Basics” you wrote! It is very well written and quite easy to follow.

While reading the book I spotted one typo and decided to send you the pull request with the correction:
I think that if there were 4 chains in the example (`chains=4`), so the following sentence “In the end, we will have three chains of …” could be corrected.

With best regards,
Vladimir